+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## 0.3.2 (Unreleased)
+## 0.4.0
+**Note: This release contains breaking changes!!**
+
+Terrahelp will now ignore stdin input if a `-file` flag is present.
+
+**The `-file` flag will no longer default to terraform.tfstate and terraform.tfstate.backup.**
+[PR-6](https://github.com/opencredo/terrahelp/pull/6)
+
+This is part of resolving [#5](https://github.com/opencredo/terrahelp/issues/5)
 
 ## 0.3.1